refactor: Group 디렉토리 구조 변경으로 깨지는 테스트 수정 (#314)
