+++ server/momo-core/src/test/java/com/momo/unit/group/application/GroupServiceTest.java
@@ -1,9 +1,7 @@
-package com.momo.unit.group.service;
+package com.momo.unit.group.application;
 
-import static com.momo.GroupFixture.getGroupCardResponse;
 import static com.momo.GroupFixture.getGroupCreateRequest;
 import static com.momo.GroupFixture.getGroupResponse;
-import static com.momo.GroupFixture.getGroupSearchConditionRequest;
 import static com.momo.GroupFixture.getGroupWithId;
 import static com.momo.UserFixture.getUserWithId;
 import static com.momo.group.domain.category.Category.EMPLOYMENT;
@@ -20,11 +18,9 @@ import com.momo.common.ServiceTest;
 import com.momo.common.exception.CustomException;
 import com.momo.common.exception.ErrorCode;
 import com.momo.group.application.GroupService;
-import com.momo.group.application.dto.response.GroupCardResponse;
 import com.momo.group.application.dto.request.GroupCreateRequest;
 import com.momo.group.application.dto.response.GroupCreateResponse;
 import com.momo.group.application.dto.response.GroupResponse;
-import com.momo.group.application.dto.request.GroupSearchConditionRequest;
 import com.momo.group.domain.Group;
 import com.momo.group.domain.participant.Participant;
 import com.momo.group.domain.repository.GroupRepository;
@@ -32,7 +28,6 @@ import com.momo.group.domain.repository.ParticipantRepository;
 import com.momo.group.domain.search.GroupSearchEngine;
 import com.momo.user.domain.User;
 import com.momo.user.domain.repository.UserRepository;
-import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -44,9 +39,6 @@ public class GroupServiceTest extends ServiceTest {
 
     @Mock
     private GroupRepository groupRepository;
-
-    @Mock
-    private GroupSearchEngine groupSearchEngine;
 
     @Mock
     private ParticipantRepository participantRepository;
@@ -67,7 +59,7 @@ public class GroupServiceTest extends ServiceTest {
         manager = getUserWithId();
         participant = getUserWithId();
         groupService = new GroupService(
-            groupRepository, groupSearchEngine, participantRepository, userRepository, s3UploadService
+            groupRepository, participantRepository, userRepository, s3UploadService
         );
     }
 
@@ -99,7 +91,7 @@ public class GroupServiceTest extends ServiceTest {
         given(userRepository.findById(any())).willReturn(of(participant));
         given(participantRepository.existsByGroupAndUser(any(), any())).willReturn(true);
 
-        groupService.updateManagerByUserId(manager, 1L, 2L);
+        groupService.updateManager(manager, 1L, 2L);
 
         verify(groupRepository).findById(any());
         verify(userRepository).findById(any());
@@ -113,7 +105,7 @@ public class GroupServiceTest extends ServiceTest {
 
         given(groupRepository.findById(any())).willReturn(of(group));
 
-        assertThatThrownBy(() -> groupService.updateManagerByUserId(participant, 1L, 2L))
+        assertThatThrownBy(() -> groupService.updateManager(participant, 1L, 2L))
             .isInstanceOf(CustomException.class)
             .hasMessage(ErrorCode.GROUP_MANAGER_AUTHORIZED.getMessage());
     }
@@ -126,7 +118,7 @@ public class GroupServiceTest extends ServiceTest {
         given(userRepository.findById(any())).willReturn(of(participant));
         given(participantRepository.existsByGroupAndUser(any(), any())).willReturn(false);
 
-        assertThatThrownBy(() -> groupService.updateManagerByUserId(manager, 1L, 2L))
+        assertThatThrownBy(() -> groupService.updateManager(manager, 1L, 2L))
             .isInstanceOf(CustomException.class)
             .hasMessage(ErrorCode.GROUP_PARTICIPANT_UNAUTHORIZED.getMessage());
     }
@@ -137,7 +129,7 @@ public class GroupServiceTest extends ServiceTest {
 
         given(groupRepository.findById(any())).willReturn(of(group));
 
-        groupService.endGroupById(manager, 1L);
+        groupService.endGroup(manager, 1L);
 
         verify(groupRepository).findById(any());
         assertThat(group.isEnd()).isTrue();
@@ -149,36 +141,8 @@ public class GroupServiceTest extends ServiceTest {
 
         given(groupRepository.findById(any())).willReturn(of(group));
 
-        assertThatThrownBy(() -> groupService.endGroupById(participant, 1L))
+        assertThatThrownBy(() -> groupService.endGroup(participant, 1L))
             .isInstanceOf(CustomException.class)
             .hasMessage(ErrorCode.GROUP_MANAGER_AUTHORIZED.getMessage());
-    }
-
-    @Test
-    void 모임_검색_테스트_성공_검색_키워드_O() {
-        GroupSearchConditionRequest request = getGroupSearchConditionRequest("키워드");
-        GroupCardResponse response = getGroupCardResponse();
-
-        given(groupSearchEngine.searchByNameLikeAndCitiesAndCategories(any(), any(), any(), any(), any()))
-            .willReturn(List.of(response));
-
-        List<GroupCardResponse> expected = groupService.findPageBySearchConditionV2(manager, request);
-
-        verify(groupSearchEngine).searchByNameLikeAndCitiesAndCategories(any(), any(), any(), any(), any());
-        assertThat(expected.size()).isEqualTo(1);
-    }
-
-    @Test
-    void 모임_검색_테스트_성공_검색_키워드_X() {
-        GroupSearchConditionRequest request = getGroupSearchConditionRequest(null);
-        GroupCardResponse response = getGroupCardResponse();
-
-        given(groupRepository.findAllByCitiesAndCategoriesOrderByCreatedDateDesc(any(), any(), any(), any()))
-            .willReturn(List.of(response));
-
-        List<GroupCardResponse> expected = groupService.findPageBySearchConditionV2(manager, request);
-
-        verify(groupRepository).findAllByCitiesAndCategoriesOrderByCreatedDateDesc(any(), any(), any(), any());
-        assertThat(expected.size()).isEqualTo(1);
     }
 }
